다듬기 전, 확률형 신청까지 만들긴 함. 함수로 모듈화 하자

+++ eventservice/src/main/java/com/chatbot/eventservice/dto/EventOutputDto.java
@@ -23,7 +23,7 @@ public class EventOutputDto {
 	private int orderCount;
 	
 	private int eventOrderCount;
-	private int dDateCount;
+	private Long dDateCount;
 	
 	private Map<String,String> resultInfo = new HashMap<String,String>();
