+++ eventservice/src/main/java/com/chatbot/eventservice/dto/EventSetupOutputDto.java
@@ -22,7 +22,7 @@ public class EventSetupOutputDto {
 	private DateType dateType; //이벤트 명 중복 date 관리(ALL - 중복 조건 없음, MIN - 분단위, HOURS - 시간 단위, DAY - 일 단위 ~~ 중복 허용)
 	private RewardType rewardType;
 
-	private LinkedHashMap<String, Integer> rewardInfo = new LinkedHashMap<>();
+	private LinkedHashMap<String, Double> rewardInfo = new LinkedHashMap<>();
 	private HashMap<String,HashMap<String,String>> resultInfo = new HashMap<>();
 
 	private List<String> targetClnn;
